changes/refactor/fix: updating order service, removing problems with not getting orders, removing build problems, changing db context model, updating contracts, refactoring

+++ src/Stroytorg.Domain/Data/Entities/OrderMaterialMap.cs
@@ -11,11 +11,11 @@ public class OrderMaterialMap : BaseEntity<int>
     [Required]
     public int MaterialId { get; set; }
 
-    public double TotalMaterialAmount { get; set; }
+    public decimal TotalMaterialAmount { get; set; }
 
-    public double UnitPrice { get; set; }
+    public decimal UnitPrice { get; set; }
 
-    public double? TotalMaterialWeight { get; set; }
+    public decimal? TotalMaterialWeight { get; set; }
 
     public virtual Order? Order { get; set; }
